+++ admin_app/src/app/dashboard/components/tempData.ts
@@ -1,0 +1,67 @@
+const topics = [
+  {
+    topic: "Topic 1",
+    count: 10,
+  },
+  {
+    topic: "Topic 2",
+    count: 20,
+  },
+  {
+    topic: "Topic 3",
+    count: 30,
+  },
+  {
+    topic: "Topic 4",
+    count: 40,
+  },
+  {
+    topic: "Topic 5",
+    count: 50,
+  },
+  {
+    topic: "Topic 6",
+    count: 60,
+  },
+  {
+    topic: "Topic 7",
+    count: 70,
+  },
+  {
+    topic: "Topic 8",
+    count: 80,
+  },
+  {
+    topic: "Topic 9",
+    count: 90,
+  },
+  {
+    topic: "Topic 10",
+    count: 100,
+  },
+];
+
+const queries = [
+  {
+    query: "What is the capital of France?",
+    datetime_utc: "2021-09-01T00:00:00Z",
+  },
+  {
+    query: "What is the capital of Germany?",
+    datetime_utc: "2021-09-01T00:00:00Z",
+  },
+  {
+    query: "What is the capital of Italy?",
+    datetime_utc: "2021-09-01T00:00:00Z",
+  },
+  {
+    query: "What is the capital of Spain?",
+    datetime_utc: "2021-09-01T00:00:00Z",
+  },
+  {
+    query: "What is the capital of Portugal?",
+    datetime_utc: "2021-09-01T00:00:00Z",
+  },
+];
+
+export { topics, queries };
